+++ src/shared/utils/orderCalculations.js
@@ -81,23 +81,23 @@ export const calculateOrderTotal = (order) => {
   if (order.furnitureData?.groups) {
     order.furnitureData.groups.forEach(group => {
       // Material costs
-      const materialTotal = (parseFloat(group.materialPrice) || 0) * (parseInt(group.materialQnty) || 0);
+      const materialTotal = (parseFloat(group.materialPrice) || 0) * (parseFloat(group.materialQnty) || 0);
       itemsSubtotal += materialTotal;
       taxableAmount += materialTotal; // Materials are taxable
       
       // Labour costs (not taxable)
-      itemsSubtotal += (parseFloat(group.labourPrice) || 0) * (parseInt(group.labourQnty) || 0);
+      itemsSubtotal += (parseFloat(group.labourPrice) || 0) * (parseFloat(group.labourQnty) || 0);
       
       // Foam costs (if enabled) - taxable
       if (group.foamEnabled || group.foamPrice > 0) {
-        const foamTotal = (parseFloat(group.foamPrice) || 0) * (parseInt(group.foamQnty) || 0);
+        const foamTotal = (parseFloat(group.foamPrice) || 0) * (parseFloat(group.foamQnty) || 0);
         itemsSubtotal += foamTotal;
         taxableAmount += foamTotal; // Foam is taxable
       }
       
       // Painting costs (if enabled) - not taxable (labour)
       if (group.paintingEnabled || group.paintingLabour > 0) {
-        const paintingTotal = (parseFloat(group.paintingLabour) || 0) * (parseInt(group.paintingQnty) || 0);
+        const paintingTotal = (parseFloat(group.paintingLabour) || 0) * (parseFloat(group.paintingQnty) || 0);
         itemsSubtotal += paintingTotal;
         // Painting is labour, so not taxable
       }
@@ -137,19 +137,19 @@ export const calculateOrderTotalWithoutTax = (order) => {
   if (order.furnitureData?.groups) {
     order.furnitureData.groups.forEach(group => {
       // Material costs
-      total += (parseFloat(group.materialPrice) || 0) * (parseInt(group.materialQnty) || 0);
+      total += (parseFloat(group.materialPrice) || 0) * (parseFloat(group.materialQnty) || 0);
       
       // Labour costs  
-      total += (parseFloat(group.labourPrice) || 0) * (parseInt(group.labourQnty) || 0);
+      total += (parseFloat(group.labourPrice) || 0) * (parseFloat(group.labourQnty) || 0);
       
       // Foam costs (if enabled)
       if (group.foamEnabled || group.foamPrice > 0) {
-        total += (parseFloat(group.foamPrice) || 0) * (parseInt(group.foamQnty) || 0);
+        total += (parseFloat(group.foamPrice) || 0) * (parseFloat(group.foamQnty) || 0);
       }
       
       // Painting costs (if enabled)
       if (group.paintingEnabled || group.paintingLabour > 0) {
-        total += (parseFloat(group.paintingLabour) || 0) * (parseInt(group.paintingQnty) || 0);
+        total += (parseFloat(group.paintingLabour) || 0) * (parseFloat(group.paintingQnty) || 0);
       }
       
       // Legacy support for old field names
@@ -178,11 +178,11 @@ export const calculateOrderTax = (order) => {
   if (order.furnitureData?.groups) {
     order.furnitureData.groups.forEach(group => {
       // Materials are taxable
-      taxableAmount += (parseFloat(group.materialPrice) || 0) * (parseInt(group.materialQnty) || 0);
+      taxableAmount += (parseFloat(group.materialPrice) || 0) * (parseFloat(group.materialQnty) || 0);
       
       // Foam is taxable
       if (group.foamEnabled || group.foamPrice > 0) {
-        taxableAmount += (parseFloat(group.foamPrice) || 0) * (parseInt(group.foamQnty) || 0);
+        taxableAmount += (parseFloat(group.foamPrice) || 0) * (parseFloat(group.foamQnty) || 0);
       }
     });
   }
@@ -206,19 +206,19 @@ export const getOrderCostBreakdown = (order) => {
   if (order.furnitureData?.groups) {
     order.furnitureData.groups.forEach(group => {
       // Material costs
-      breakdown.material += (parseFloat(group.materialPrice) || 0) * (parseInt(group.materialQnty) || 0);
+      breakdown.material += (parseFloat(group.materialPrice) || 0) * (parseFloat(group.materialQnty) || 0);
       
       // Labour costs
-      breakdown.labour += (parseFloat(group.labourPrice) || 0) * (parseInt(group.labourQnty) || 0);
+      breakdown.labour += (parseFloat(group.labourPrice) || 0) * (parseFloat(group.labourQnty) || 0);
       
       // Foam costs
       if (group.foamEnabled || group.foamPrice > 0) {
-        breakdown.foam += (parseFloat(group.foamPrice) || 0) * (parseInt(group.foamQnty) || 0);
+        breakdown.foam += (parseFloat(group.foamPrice) || 0) * (parseFloat(group.foamQnty) || 0);
       }
       
       // Painting costs
       if (group.paintingEnabled || group.paintingLabour > 0) {
-        breakdown.painting += (parseFloat(group.paintingLabour) || 0) * (parseInt(group.paintingQnty) || 0);
+        breakdown.painting += (parseFloat(group.paintingLabour) || 0) * (parseFloat(group.paintingQnty) || 0);
       }
       
       // Legacy support
